Use the more specific label rel=source links
Instead of generic STAC derived_from rel.

We're not really sure what the difference is TBH,
as both semantically link back to a different STAC
Item that this Item is derived from.

Assumes that the label asset has key "labels"
according to a loose interpretation of the
STAC label extension spec.