Update USFIMR catalog generator to use custom merged SHP
The one on the USFIMR FTP site has bad coordinates.

This one was generated manually according to:
https://github.com/azavea/noaa-flood-mapping/issues/55#issuecomment-674954618

And uploaded to s3://usfimr-data/shp/USFIMR_merged.zip

This commit updates the build_catalog.sh script to pull
our file instead, unzip it and pass it to build_catalog.py

I switched use of /tmp to the ./data dir so that all project
files are contained within this subdir rather than leaving
large files elsewhere on the filesystem.

You'll need the awscli installed to pull the file now,
its added to requirements.txt